aircraft config and flight route changes

+++ FlightReservationSystem/FlightReservationSystem-ejb/src/java/ejb/session/stateless/FlightRouteSessionBean.java
@@ -5,6 +5,7 @@
  */
 package ejb.session.stateless;
 
+import entity.Flight;
 import entity.FlightRoute;
 import java.util.List;
 import javax.ejb.Stateless;
@@ -32,18 +33,45 @@ public class FlightRouteSessionBean implements FlightRouteSessionBeanRemote {
     }
     
     @Override
-    public List<FlightRoute> viewAllFlightRoutes() {
+    public List<FlightRoute> retrieveAllFlightRoutes() {
         
-        TypedQuery<FlightRoute> query = em.createQuery("SELECT a FROM AircraftConfiguration a", FlightRoute.class);
+        TypedQuery<FlightRoute> query = em.createQuery("SELECT f FROM FlightRoute f", FlightRoute.class);
         List<FlightRoute> flightRoutes = query.getResultList();
+        // sort by ascending order
+        flightRoutes.sort((FlightRoute fr1, FlightRoute fr2) -> fr1.getOriginAirport().getAirportName().compareTo(fr2.getOriginAirport().getAirportName()));
+        // insert return flights if any
+        int numOfRoutes = flightRoutes.size();
+        for (int i = 0; i < numOfRoutes; i++) {
+            FlightRoute route1 = flightRoutes.get(i);
+            if (route1.getReturnFlightRoute() != null) { //  return flight route exists
+                for (int j = i+1; j < numOfRoutes; j++) {
+                    if (route1.getReturnFlightRoute().equals(flightRoutes.get(j))) {
+                        FlightRoute returnRoute = flightRoutes.get(j);
+                        flightRoutes.remove(j);
+                        flightRoutes.add(i+1, returnRoute);
+                    }
+                }
+            }
+        }
         
         return flightRoutes;
     }
 
     @Override
-    public void deleteFlightRoute(FlightRoute flightRoute) {
+    public void deleteFlightRoute(Long flightRouteId) {
         
-        em.remove(flightRoute);
+        FlightRoute flightRoute = em.find(FlightRoute.class, flightRouteId);
+        List<Flight> flights = flightRoute.getFlights();
+        if (flights.isEmpty()) {
+            if (flightRoute.getReturnFlightRoute() != null) {
+                flightRoute.getReturnFlightRoute().setReturnFlightRoute(null);
+                // persist?
+            }
+            em.remove(flightRoute);
+        } else {
+            // disassociate with other flights?
+             flightRoute.setEnabled(false);
+        }
     }
     
 }
